removed lib/app as it is not used anymore
